file size, frontend
