--- env.example
@@ -1,43 +1,0 @@
-# Database Configuration
-DATABASE_TYPE=postgresql
-POSTGRESQL_HOST=localhost
-POSTGRESQL_PORT=5432
-POSTGRESQL_USER=postgres
-POSTGRESQL_PASSWORD=password
-POSTGRESQL_DATABASE=speech2sql
-POSTGRESQL_URL=postgresql://postgres:password@localhost:5432/speech2sql
-
-# API Keys
-UPSTAGE_API_KEY=your_upstage_api_key_here
-OPENAI_API_KEY=your_openai_api_key_here
-
-# Elasticsearch Configuration
-ELASTICSEARCH_URL=http://localhost:9200
-ELASTICSEARCH_USERNAME=elastic
-ELASTICSEARCH_PASSWORD=your_elasticsearch_password_here
-
-# Model Configuration
-WHISPER_MODEL=base  # tiny, base, small, medium, large
-SUMMARIZATION_MODEL=pegasus-large  # pegasus-large, llama2-7b
-TEXT2SQL_MODEL=text2sql-large
-
-# Audio Processing
-AUDIO_UPLOAD_PATH=./data/raw
-PROCESSED_AUDIO_PATH=./data/processed
-MAX_AUDIO_SIZE=100MB
-SUPPORTED_FORMATS=wav,mp3,m4a
-
-# Server Configuration
-HOST=0.0.0.0
-PORT=8000
-DEBUG=True
-LOG_LEVEL=INFO
-
-# File Storage
-UPLOAD_DIR=./uploads
-TEMP_DIR=./temp
-MODEL_CACHE_DIR=./data/models
-
-# Security
-SECRET_KEY=your_secret_key_here
-ACCESS_TOKEN_EXPIRE_MINUTES=30
