+++ frontend/modules/analytics.py
@@ -1,7 +1,1 @@
 # 분석 대시보드
-
-# 임시
-import streamlit as st
-
-class Analytics:
-    pass
